fix(memory): correct three defects found by adversarial review
An adversarial review of this branch raised eight findings across four lenses;
six survived independent refutation, and those six are four distinct defects.
Three are fixed here. The two concerning `murmly doctor` are design questions
rather than mistakes and are being taken separately.

The synthesis countdown was armed after `_session_closed` released
`_speech_session_lock`, and after a drain that waits on the writer thread. A
declaration arriving in that window registers its own session and cancels the
countdown, and the stale arm then runs against a session that has never been
idle -- so an opted-in synthesis period could drop the ONNX session underneath a
live sender, costing silence at the front of its next reply. The code already
knew this: the failed-start arm site arms inside its identity branch and its
comment names exactly this hazard. This site now does the same. Reproduced 32
times in 800 iterations against real session objects over socketpairs before the
fix, with no patching.

`_idle_period` tested `int(value) == 0`, which truncates rather than compares, so
0.5, 0.99 and -0.9 all read as zero and switched idle release off. The spec
requires a value outside the 30-86400 bounds to fall back to that setting's own
default, and for `[stt]`, whose default is 300, this silently disabled a feature
that ships enabled -- the same inversion the helper was written to prevent, one
layer further in. Only an exact zero now means never.

The README claimed Murmly loads the transcription model on the first recording,
which is false under `lazy_load_model = false`, documented two sections above it.

Both code fixes are pinned by tests that fail against the previous behaviour: the
countdown one asserts the arm happens while the session lock is held, since the
interleaving is a race a test cannot schedule, and the config one covers the
fractional band the existing out-of-range test never reached.

759 tests pass, 2 skipped.

+++ src/murmly/config.py
@@ -380,7 +380,12 @@ def _idle_period(value: object, default: int) -> int:
     """
     if value is not None:
         try:
-            if int(value) == 0:
+            # Compared, not truncated. `int(value)` reads 0.5 and -0.9 as zero,
+            # so a mistyped fractional period switched release off instead of
+            # falling back -- and for `[stt]`, whose default is 300, that
+            # silently disabled a feature that ships on. Only an exact zero
+            # means never; everything else outside the bounds falls through.
+            if float(value) == 0:
                 return 0
         except (TypeError, ValueError):
             pass
